+++ src/blocks/graph/__quantity/graph__quantity.css
@@ -9,53 +9,3 @@
   font-weight: bold;
   padding-left: 6px;
 }
-
-/*.graph__quantity:first-of-type {
-  width: 20%;
-}
-
-.graph__quantity:nth-of-type(2) {
-  width: 25%;
-}
-
-.graph__quantity:nth-of-type(3) {
-  width: 15%;
-}
-.graph__quantity:nth-of-type(4) {
-  width: 50%;
-}
-.graph__quantity:nth-of-type(5) {
-  width: 70%;
-}
-.graph__quantity:nth-of-type(6) {
-  width: 1fr;
-}
-.graph__quantity:nth-of-type(7) {
-  width: 45%;
-}
-
-@media (min-width: 1440px){
-.graph__quantity:first-of-type {
-  width: 15%;
-}
-
-.graph__quantity:nth-of-type(2) {
-  width: 18%;
-}
-
-.graph__quantity:nth-of-type(3) {
-  width: 11%;
-}
-.graph__quantity:nth-of-type(4) {
-  width: 39%;
-}
-.graph__quantity:nth-of-type(5) {
-  width: 45%;
-}
-.graph__quantity:nth-of-type(6) {
-  width: 73%;
-}
-.graph__quantity:nth-of-type(7) {
-  width: 34%;
-}
-}*/
